Move ContextConfigurationWithTestChannel class (#342)
