+++ src/test/java/org/gridsuite/study/server/StudyControllerDynamicSimulationTest.java
@@ -18,7 +18,6 @@ import org.gridsuite.study.server.notification.NotificationService;
 import org.gridsuite.study.server.service.NetworkModificationTreeService;
 import org.gridsuite.study.server.service.StudyService;
 import org.gridsuite.study.server.service.client.util.UrlUtil;
-import org.gridsuite.study.server.utils.ContextConfigurationWithTestChannel;
 import org.gridsuite.study.server.utils.elasticsearch.DisableElasticsearch;
 import org.gridsuite.study.server.utils.TestUtils;
 import org.junit.After;
